Use a long int for iSamplesToRead.

+++ src/RageSoundReader_ChannelSplit.cpp
@@ -216,8 +216,9 @@ int RageSoundSplitterImpl::ReadBuffer()
 	if( iFramesToRead <= 0 )
 		return 1; // requested data already buffered
 
-	int iSamplesToRead = iFramesToRead * m_pSource->GetNumChannels();
-	int iOldSizeSamples = m_sBuffer.size();
+	// Resize the buffer to accommodate the new frames.
+	long int iSamplesToRead = iFramesToRead * m_pSource->GetNumChannels();
+	size_t iOldSizeSamples = m_sBuffer.size();
 	m_sBuffer.resize( iOldSizeSamples + iSamplesToRead );
 	int iGotFrames = m_pSource->Read( &m_sBuffer[0] + iOldSizeSamples, iFramesToRead );
 	if( iGotFrames < 0 )
